pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "imganalyzer"
version = "0.1.0"
description = "CLI tool to analyze images and camera RAW files with EXIF, technical, and AI analysis — outputs Lightroom-compatible XMP sidecars"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["image", "photography", "raw", "exif", "xmp", "lightroom", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "typer[all]>=0.12",
    "Pillow>=10.0",
    "rawpy>=0.21",
    "exifread>=3.0",
    "piexif>=1.1",
    "numpy>=1.24",
    "scikit-image>=0.22",
    "rich>=13.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "lxml>=5.0",
]

[project.optional-dependencies]
local-ai = [
    "transformers>=4.40",
    "torch>=2.0",
    "accelerate>=0.29",
]
openai = ["openai>=1.30"]
anthropic = ["anthropic>=0.28"]
google = ["google-cloud-vision>=3.7"]
all-ai = [
    "imganalyzer[local-ai]",
    "imganalyzer[openai]",
    "imganalyzer[anthropic]",
    "imganalyzer[google]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
imganalyzer = "imganalyzer.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["imganalyzer*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
